Multi source multi sink, unit tested, manual component tested
